# apps/cms/openapi.yaml
openapi: 3.1.0
info:
  title: LumiBase CMS API
  version: 0.1.0
  description: |
    Edge-native Headless CMS API. See `docs/api/hono-api-spec.md` for the full
    surface; this file is the machine-readable contract that ships with the
    repo. Endpoints land incrementally — Phase 0 covers auth/health only.
servers:
  - url: https://api.lumibase.dev/api/v1
    description: Production
  - url: http://127.0.0.1:1989/api/v1
    description: Local wrangler dev
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    SiteHeader:
      name: X-Lumi-Site
      in: header
      required: true
      schema: { type: string }
      description: Active tenant id. Required for every authenticated endpoint.
  schemas:
    Error:
      type: object
      properties:
        errors:
          type: array
          items:
            type: object
            required: [code, message]
            properties:
              code: { type: string }
              message: { type: string }
              path:
                type: array
                items: { type: string }
              trace: { type: object, additionalProperties: true }
...
    ShareResource:
      type: object
      required: [id, siteId, collection, itemId, roleId, usedCount]
      properties:
        id: { type: string }
        siteId: { type: string }
        collection: { type: string }
        itemId: { type: string }
        roleId: { type: string }
        validFrom: { type: string, format: date-time, nullable: true }
        validUntil: { type: string, format: date-time, nullable: true }
        maxUses: { type: integer, nullable: true }
        usedCount: { type: integer }
        revokedAt: { type: string, format: date-time, nullable: true }
        revokedBy: { type: string, nullable: true }
        createdBy: { type: string, nullable: true }
        lastUsedAt: { type: string, format: date-time, nullable: true }
        createdAt: { type: string, format: date-time }
    ShareSecretResult:
      allOf:
        - $ref: '#/components/schemas/ShareResource'
        - type: object
          required: [token, url]
          properties:
            token:
              type: string
              description: Plaintext share token returned only on create.
            url: { type: string }
    ShareCreateRequest:
      type: object
      required: [collection, itemId, roleId]
      properties:
        collection: { type: string, minLength: 1 }
        itemId: { type: string, minLength: 1 }
        roleId: { type: string, minLength: 1 }
        password: { type: string, minLength: 1 }
        validFrom: { type: string, format: date-time, nullable: true }
        validUntil: { type: string, format: date-time, nullable: true }
        maxUses: { type: integer, minimum: 1, nullable: true }
    # --- AI (POST-GA1) ---
...
    AgentRun:
      type: object
      properties:
        id: { type: string }
        goalId: { type: string }
        siteId: { type: string }
        agentName: { type: string }
        provider: { type: string }
        model: { type: string }
        status: { type: string }
        budget: { type: object, additionalProperties: true }
        risk: { type: string, enum: [safe, review_required, dangerous, blocked] }
        metrics: { type: object, additionalProperties: true }
        error: { type: string, nullable: true }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...
    AgentArtifact:
      type: object
      properties:
        id: { type: string }
        runId: { type: string }
        siteId: { type: string }
        type: { type: string, enum: [schema_diff, page_spec, component_spec, seed_data, api_spec, prompt, migration] }
        target: { type: string, nullable: true }
        title: { type: string }
        contentRef: { type: string, nullable: true }
        content: { type: object, additionalProperties: true }
        hash: { type: string }
        version: { type: integer }
        status: { type: string }
        metadata: { type: object, additionalProperties: true }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...
    AgentMemoryContext:
      type: object
      properties:
        siteId: { type: string }
        memories:
          type: array
          items: { type: object, additionalProperties: true }
        recentRuns:
          type: array
          items: { $ref: '#/components/schemas/AgentRun' }
        approvedArtifacts:
          type: array
          items: { $ref: '#/components/schemas/AgentArtifact' }
    # --- Flows (POST-GA2) ---
...
paths:
...
  /shares:
    parameters: [{ $ref: '#/components/parameters/SiteHeader' }]
    post:
      summary: Create a public share link for an item
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ShareCreateRequest' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: '#/components/schemas/ShareSecretResult' }
  /shares/{token}:
    get:
      summary: Read a shared item through the share role permission
      security: []
      parameters:
        - name: token
          in: path
          required: true
          schema: { type: string }
        - name: X-Lumi-Share-Password
          in: header
          required: false
          schema: { type: string }
        - name: password
          in: query
          required: false
          schema: { type: string }
      responses:
        '200':
          description: Shared item payload, field-masked by the share role
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { type: object, additionalProperties: true }
        '401':
          description: Password required or share expired
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '403':
          description: Share revoked, not yet valid, max uses reached, or role denied
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /shares/{id}/revoke:
    parameters:
      - $ref: '#/components/parameters/SiteHeader'
      - { name: id, in: path, required: true, schema: { type: string } }
    post:
      summary: Revoke a share link
      responses:
        '200':
          description: Revoked
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: '#/components/schemas/ShareResource' }
...
  /agent/memory:
    get:
      summary: Build agent memory context
      parameters:
        - $ref: '#/components/parameters/SiteHeader'
        - { name: scope, in: query, required: false, schema: { type: string } }
        - { name: scopeId, in: query, required: false, schema: { type: string } }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: '#/components/schemas/AgentMemoryContext' }
    post:
      summary: Write agent memory
      parameters: [{ $ref: '#/components/parameters/SiteHeader' }]
      responses:
        '201': { description: Created }
...